hurd: Update struct statfs according to struct statvfs
	* sysdeps/mach/hurd/bits/statfs.h (struct statfs): Make f_bsize,
	f_namemax, f_frsize, and f_flag fields unsigned long int instead of
	unsigned int.
	(struct statfs64): Likewise.

+++ sysdeps/mach/hurd/bits/statfs.h
@@ -37,7 +37,7 @@
 struct statfs
   {
     unsigned int f_type;
-    unsigned int f_bsize;
+    unsigned long int f_bsize;
 #ifndef __USE_FILE_OFFSET64
     __fsblkcnt_t f_blocks;
     __fsblkcnt_t f_bfree;
@@ -52,14 +52,14 @@ struct statfs
     __fsblkcnt64_t f_ffree;
 #endif
     __fsid_t f_fsid;
-    unsigned int f_namelen;
+    unsigned long int f_namelen;
 #ifndef __USE_FILE_OFFSET64
     __fsfilcnt_t f_favail;
 #else
     __fsfilcnt64_t f_favail;
 #endif
-    unsigned int f_frsize;
-    unsigned int f_flag;
+    unsigned long int f_frsize;
+    unsigned long int f_flag;
     unsigned int f_spare[3];
   };
 
@@ -67,17 +67,17 @@ struct statfs
 struct statfs64
   {
     unsigned int f_type;
-    unsigned int f_bsize;
+    unsigned long int f_bsize;
     __fsblkcnt64_t f_blocks;
     __fsblkcnt64_t f_bfree;
     __fsblkcnt64_t f_bavail;
     __fsblkcnt64_t f_files;
     __fsblkcnt64_t f_ffree;
     __fsid_t f_fsid;
-    unsigned int f_namelen;
+    unsigned long int f_namelen;
     __fsfilcnt64_t f_favail;
-    unsigned int f_frsize;
-    unsigned int f_flag;
+    unsigned long int f_frsize;
+    unsigned long int f_flag;
     unsigned int f_spare[3];
   };
 #endif
